Move lib files into ./lib
