+++ lib/spotlight.js
